Namespace Input and deprecate it in favor of the framework package (#16939)
* Apply patch #12532, move to the fw package classes

* Namespace input classes

* Change application classes

* Compile correct class name

* Restore unit test
