+++ libraries/src/Joomla/CMS/Application/CliApplication.php
@@ -11,6 +11,8 @@ namespace Joomla\CMS\Application;
 defined('JPATH_PLATFORM') or die;
 
 use Joomla\Application\Cli\CliOutput;
+use Joomla\CMS\Input\Cli;
+use Joomla\CMS\Input\Input;
 use Joomla\Registry\Registry;
 
 /**
@@ -36,21 +38,21 @@ class CliApplication extends BaseApplication
 	/**
 	 * Class constructor.
 	 *
-	 * @param   \JInputCli         $input       An optional argument to provide dependency injection for the application's
-	 *                                         input object.  If the argument is a \JInputCli object that object will become
-	 *                                         the application's input object, otherwise a default input object is created.
+	 * @param   Cli                $input       An optional argument to provide dependency injection for the application's
+	 *                                          input object.  If the argument is a \JInputCli object that object will become
+	 *                                          the application's input object, otherwise a default input object is created.
 	 * @param   Registry           $config      An optional argument to provide dependency injection for the application's
-	 *                                         config object.  If the argument is a Registry object that object will become
-	 *                                         the application's config object, otherwise a default config object is created.
+	 *                                          config object.  If the argument is a Registry object that object will become
+	 *                                          the application's config object, otherwise a default config object is created.
 	 * @param   \JEventDispatcher  $dispatcher  An optional argument to provide dependency injection for the application's
-	 *                                         event dispatcher.  If the argument is a \JEventDispatcher object that object will become
-	 *                                         the application's event dispatcher, if it is null then the default event dispatcher
-	 *                                         will be created based on the application's loadDispatcher() method.
+	 *                                          event dispatcher.  If the argument is a \JEventDispatcher object that object will become
+	 *                                          the application's event dispatcher, if it is null then the default event dispatcher
+	 *                                          will be created based on the application's loadDispatcher() method.
 	 *
 	 * @see     BaseApplication::loadDispatcher()
 	 * @since   11.1
 	 */
-	public function __construct(\JInputCli $input = null, Registry $config = null, \JEventDispatcher $dispatcher = null)
+	public function __construct(Cli $input = null, Registry $config = null, \JEventDispatcher $dispatcher = null)
 	{
 		// Close the application if we are not executed from the command line.
 		// @codeCoverageIgnoreStart
@@ -61,16 +63,16 @@ class CliApplication extends BaseApplication
 		// @codeCoverageIgnoreEnd
 
 		// If an input object is given use it.
-		if ($input instanceof \JInput)
+		if ($input instanceof Input)
 		{
 			$this->input = $input;
 		}
 		// Create the input based on the application logic.
 		else
 		{
-			if (class_exists('\JInput'))
+			if (class_exists('\\Joomla\\CMS\\Input\\Cli'))
 			{
-				$this->input = new \JInputCli;
+				$this->input = new Cli;
 			}
 		}
