+++ libraries/src/Joomla/CMS/Application/WebApplication.php
@@ -10,6 +10,7 @@ namespace Joomla\CMS\Application;
 
 defined('JPATH_PLATFORM') or die;
 
+use Joomla\CMS\Input\Input;
 use Joomla\Registry\Registry;
 use Joomla\String\StringHelper;
 
@@ -125,29 +126,29 @@ class WebApplication extends BaseApplication
 	/**
 	 * Class constructor.
 	 *
-	 * @param   \JInput                 $input   An optional argument to provide dependency injection for the application's
-	 *                                          input object.  If the argument is a \JInput object that object will become
-	 *                                          the application's input object, otherwise a default input object is created.
+	 * @param   Input                   $input   An optional argument to provide dependency injection for the application's
+	 *                                           input object.  If the argument is a \JInput object that object will become
+	 *                                           the application's input object, otherwise a default input object is created.
 	 * @param   Registry                $config  An optional argument to provide dependency injection for the application's
-	 *                                          config object.  If the argument is a Registry object that object will become
-	 *                                          the application's config object, otherwise a default config object is created.
+	 *                                           config object.  If the argument is a Registry object that object will become
+	 *                                           the application's config object, otherwise a default config object is created.
 	 * @param   \JApplicationWebClient  $client  An optional argument to provide dependency injection for the application's
-	 *                                          client object.  If the argument is a \JApplicationWebClient object that object will become
-	 *                                          the application's client object, otherwise a default client object is created.
+	 *                                           client object.  If the argument is a \JApplicationWebClient object that object will become
+	 *                                           the application's client object, otherwise a default client object is created.
 	 *
 	 * @since   11.3
 	 */
-	public function __construct(\JInput $input = null, Registry $config = null, \JApplicationWebClient $client = null)
+	public function __construct(Input $input = null, Registry $config = null, \JApplicationWebClient $client = null)
 	{
 		// If an input object is given use it.
-		if ($input instanceof \JInput)
+		if ($input instanceof Input)
 		{
 			$this->input = $input;
 		}
 		// Create the input based on the application logic.
 		else
 		{
-			$this->input = new \JInput;
+			$this->input = new Input;
 		}
 
 		// If a config object is given use it.
