+++ libraries/src/Joomla/CMS/Input/Cli.php
@@ -1,26 +1,31 @@
 <?php
 /**
- * @package     Joomla.Platform
- * @subpackage  Input
+ * Joomla! Content Management System
  *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE
+ * @copyright  Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
  */
 
+namespace Joomla\CMS\Input;
+
 defined('JPATH_PLATFORM') or die;
+
+use Joomla\CMS\Filter\InputFilter;
 
 /**
  * Joomla! Input CLI Class
  *
- * @since  11.1
+ * @since       11.1
+ * @deprecated  5.0  Use Joomla\Input\Cli instead
  */
-class JInputCli extends JInput
+class Cli extends Input
 {
 	/**
 	 * The executable that was called to run the CLI script.
 	 *
 	 * @var    string
 	 * @since  11.1
+	 * @deprecated  5.0  Use Joomla\Input\Cli instead
 	 */
 	public $executable;
 
@@ -30,6 +35,7 @@ class JInputCli extends JInput
 	 *
 	 * @var    array
 	 * @since  11.1
+	 * @deprecated  5.0  Use Joomla\Input\Cli instead
 	 */
 	public $args = array();
 
@@ -40,6 +46,7 @@ class JInputCli extends JInput
 	 * @param   array  $options  Array of configuration parameters (Optional)
 	 *
 	 * @since   11.1
+	 * @deprecated  5.0  Use Joomla\Input\Cli instead
 	 */
 	public function __construct(array $source = null, array $options = array())
 	{
@@ -49,7 +56,7 @@ class JInputCli extends JInput
 		}
 		else
 		{
-			$this->filter = JFilterInput::getInstance();
+			$this->filter = InputFilter::getInstance();
 		}
 
 		// Get the command line options
@@ -65,6 +72,7 @@ class JInputCli extends JInput
 	 * @return  string  The serialized input.
 	 *
 	 * @since   12.1
+	 * @deprecated  5.0  Use Joomla\Input\Cli instead
 	 */
 	public function serialize()
 	{
@@ -85,9 +93,10 @@ class JInputCli extends JInput
 	 *
 	 * @param   string  $input  The serialized input.
 	 *
-	 * @return  JInput  The input object.
+	 * @return  Input  The input object.
 	 *
 	 * @since   12.1
+	 * @deprecated  5.0  Use Joomla\Input\Cli instead
 	 */
 	public function unserialize($input)
 	{
@@ -101,7 +110,7 @@ class JInputCli extends JInput
 		}
 		else
 		{
-			$this->filter = JFilterInput::getInstance();
+			$this->filter = InputFilter::getInstance();
 		}
 	}
 
@@ -113,6 +122,7 @@ class JInputCli extends JInput
 	 * @return  void
 	 *
 	 * @since   11.1
+	 * @deprecated  5.0  Use Joomla\Input\Cli instead
 	 */
 	protected function parseArguments()
 	{
