+++ libraries/src/Joomla/CMS/Input/Files.php
@@ -1,26 +1,31 @@
 <?php
 /**
- * @package     Joomla.Platform
- * @subpackage  Input
+ * Joomla! Content Management System
  *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE
+ * @copyright  Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
  */
 
+namespace Joomla\CMS\Input;
+
 defined('JPATH_PLATFORM') or die;
+
+use Joomla\CMS\Filter\InputFilter;
 
 /**
  * Joomla! Input Files Class
  *
- * @since  11.1
+ * @since       11.1
+ * @deprecated  5.0  Use Joomla\Input\Files instead
  */
-class JInputFiles extends JInput
+class Files extends Input
 {
 	/**
 	 * The pivoted data from a $_FILES or compatible array.
 	 *
 	 * @var    array
 	 * @since  11.1
+	 * @deprecated  5.0  Use Joomla\Input\Files instead
 	 */
 	protected $decodedData = array();
 
@@ -32,6 +37,7 @@ class JInputFiles extends JInput
 	 *                           filter : a custom JFilterInput object.
 	 *
 	 * @since   12.1
+	 * @deprecated  5.0  Use Joomla\Input\Files instead
 	 */
 	public function __construct(array $source = null, array $options = array())
 	{
@@ -41,7 +47,7 @@ class JInputFiles extends JInput
 		}
 		else
 		{
-			$this->filter = JFilterInput::getInstance();
+			$this->filter = InputFilter::getInstance();
 		}
 
 		// Set the data source.
@@ -62,6 +68,7 @@ class JInputFiles extends JInput
 	 *
 	 * @see     JFilterInput::clean()
 	 * @since   11.1
+	 * @deprecated  5.0  Use Joomla\Input\Files instead
 	 */
 	public function get($name, $default = null, $filter = 'cmd')
 	{
@@ -80,7 +87,7 @@ class JInputFiles extends JInput
 			// Prevent returning an unsafe file unless speciffically requested
 			if ($filter != 'raw')
 			{
-				$isSafe = JFilterInput::isSafeFile($results);
+				$isSafe = InputFilter::isSafeFile($results);
 
 				if (!$isSafe)
 				{
@@ -102,6 +109,7 @@ class JInputFiles extends JInput
 	 * @return  array
 	 *
 	 * @since   11.1
+	 * @deprecated  5.0  Use Joomla\Input\Files instead
 	 */
 	protected function decodeData(array $data)
 	{
@@ -129,6 +137,7 @@ class JInputFiles extends JInput
 	 * @return  void
 	 *
 	 * @since   11.1
+	 * @deprecated  5.0  Use Joomla\Input\Files instead
 	 */
 	public function set($name, $value)
 	{
