+++ libraries/src/Joomla/CMS/Input/Json.php
@@ -1,11 +1,14 @@
 <?php
 /**
- * @package     Joomla.Platform
- * @subpackage  Input
+ * Joomla! Content Management System
  *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE
+ * @copyright  Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
  */
+
+namespace Joomla\CMS\Input;
+
+use Joomla\CMS\Filter\InputFilter;
 
 defined('JPATH_PLATFORM') or die;
 
@@ -15,13 +18,15 @@ defined('JPATH_PLATFORM') or die;
  * This class decodes a JSON string from the raw request data and makes it available via
  * the standard JInput interface.
  *
- * @since  12.2
+ * @since       12.2
+ * @deprecated  5.0  Use Joomla\Input\Json instead
  */
-class JInputJSON extends JInput
+class Json extends Input
 {
 	/**
 	 * @var    string  The raw JSON string from the request.
 	 * @since  12.2
+	 * @deprecated  5.0  Use Joomla\Input\Json instead
 	 */
 	private $_raw;
 
@@ -32,6 +37,7 @@ class JInputJSON extends JInput
 	 * @param   array  $options  Array of configuration parameters (Optional)
 	 *
 	 * @since   12.2
+	 * @deprecated  5.0  Use Joomla\Input\Json instead
 	 */
 	public function __construct(array $source = null, array $options = array())
 	{
@@ -41,7 +47,7 @@ class JInputJSON extends JInput
 		}
 		else
 		{
-			$this->filter = JFilterInput::getInstance();
+			$this->filter = InputFilter::getInstance();
 		}
 
 		if (is_null($source))
@@ -64,6 +70,7 @@ class JInputJSON extends JInput
 	 * @return  string  The raw JSON string from the request.
 	 *
 	 * @since   12.2
+	 * @deprecated  5.0  Use Joomla\Input\Json instead
 	 */
 	public function getRaw()
 	{
