+++ libraries/src/Joomla/CMS/Session/Session.php
@@ -11,6 +11,7 @@ namespace Joomla\CMS\Session;
 defined('JPATH_PLATFORM') or die;
 
 use Joomla\CMS\Application\ApplicationHelper;
+use Joomla\CMS\Input\Input;
 use Joomla\CMS\User\UserHelper;
 
 /**
@@ -450,14 +451,14 @@ class Session implements \IteratorAggregate
 	/**
 	 * Check whether this session is currently created
 	 *
-	 * @param   \JInput            $input       \JInput object for the session to use.
+	 * @param   Input              $input       Input object for the session to use.
 	 * @param   \JEventDispatcher  $dispatcher  Dispatcher object for the session to use.
 	 *
 	 * @return  void
 	 *
 	 * @since   12.2
 	 */
-	public function initialise(\JInput $input, \JEventDispatcher $dispatcher = null)
+	public function initialise(Input $input, \JEventDispatcher $dispatcher = null)
 	{
 		// With the introduction of the handler class this variable is no longer required
 		// however we keep setting it for b/c
